requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
databases==0.9.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysqlclient==2.2.4
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1